edit update habit to update streak, refactor streak functions in database.js
